requirements.txt
flask
PyNaCl
requests
flask[async]
pymongo
pymongo[srv]
python-dotenv